🎨 style: 修复代码评审发现的问题
Clippy 警告修复：
- 移除不必要的 return 语句
- 合并嵌套的 if 语句 (collapsible_if)
- 使用 div_ceil() 替代手动计算
- 移除不必要的类型转换 (as f32)
- 使用 matches! 宏替代返回 bool 的 match
- 使用 RangeInclusive::contains() 替代手动比较
- 使用 #[derive(Default)] 替代手动 impl
- 为 ECS 系统函数添加 #[allow(clippy::too_many_arguments)]
- 为复杂类型添加 #[allow(clippy::type_complexity)]

代码改进：
- tokio_tasks: 改善 panic 消息并添加日志
- database: 优化 add_completed_episode_async 只查询必要字段

+++ crates/picacg_app/src/systems/favorites.rs
@@ -709,7 +709,7 @@ pub fn update_favorites_content_size(
     let scale_factor = window_query
         .single()
         .ok()
-        .map(|w| w.scale_factor() as f32)
+        .map(|w| w.scale_factor())
         .unwrap_or(1.0);
 
     for (scroll_computed, mut content_info, children) in scroll_query.iter_mut() {
